Fix select sleep range continued
The sqlite date function parses a ISO date with timezone to UTC date.
For example: date('2018-03-11T00:01:00+02:00') would be parsed to '2018-03-10'.

The problem is that dates that are expected to be included in the selected range falls outside it because of the UTC adjustment.

Added additional local date fields in the database to select on, which is more clean than using substr to get the local date.